requirements.txt
fastapi
python-dotenv
supabase
python-multipart
requests
pydantic
pydantic-settings
rank_bm25
passlib[bcrypt]
bcrypt==3.2.2
python-jose
groq
